Print the name of the config files that are loaded or skipped
Add a specific task to debug configuration: config. To use it:
'manuel config geojb' or 'manuel config prod geojb'.

It will pretty print the configuration to stdout and print debugging
informations about the files that are loaded to stderr.

The a specific section of the configuration can be access with:
manuel config geojb | jq '.mapserver'

Close #1