Improve detection of the root layer to ignore it
The root layer is a "false" group that contains all other groups and
layers. We don't need it so we ignore it.

We use to ignore it only in the search but the detection could cause
problems (it was based on the name). We now rely on the parent property
of the layer and we don't add it to the layer definition.